"refactor: ApiResponse -> global 이동 "
