+++ src/main/java/com/team5/catdogeats/global/dto/ApiResponse.java
@@ -1,10 +1,12 @@
-package com.team5.catdogeats.users.dto;
-
-import lombok.*;
+package com.team5.catdogeats.global.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
+/**
+ * 전역 API 응답 포맷
+ * 모든 REST API 응답에서 사용하는 공통 응답 구조
+ */
 public record ApiResponse<T>(
         boolean success,
         String message,
@@ -13,7 +15,9 @@ public record ApiResponse<T>(
         String path,
         List<FieldError> errors
 ) {
-    // 정적 팩토리 메서드들은 그대로 유지
+    /**
+     * 성공 응답 생성
+     */
     public static <T> ApiResponse<T> success(String message, T data) {
         return new ApiResponse<>(
                 true,
@@ -25,6 +29,9 @@ public record ApiResponse<T>(
         );
     }
 
+    /**
+     * 실패 응답 생성
+     */
     public static <T> ApiResponse<T> error(String message, String path, List<FieldError> errors) {
         return new ApiResponse<>(
                 false,
@@ -36,7 +43,16 @@ public record ApiResponse<T>(
         );
     }
 
-    // FieldError도 Record로 변환
+    /**
+     * 단순 에러 응답 생성 (필드 에러 없이)
+     */
+    public static <T> ApiResponse<T> error(String message, String path) {
+        return error(message, path, null);
+    }
+
+    /**
+     * 필드 에러 정보
+     */
     public record FieldError(
             String field,
             String message
